build version helpers
